refactoring all methods was done
