--- Tests/DataStructuresPackageTests/ValuesStub.swift
@@ -1,16 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Yuri Volegov on 08.01.2024.
-//
-
-import Foundation
-
-enum ValuesStub: Int {
-	case firstValue = 2
-	case secondValue = 3
-	case fhirdValue = 5
-	case fouthValue = 9
-	case failIntTestValue = 999999
-}
